requirements.txt
langchain
langchain-chroma
langchain-community
langchain-openai
python-dotenv
langchain_core